TINKERPOP-3136 Setup embedded test suite
Moved all lambda testing to LambdaStepTest for the java-based testing.
Deprecated the "limited" tests in favor of a newly added "embedded" test suite.
Removed all lambda tests from gherkin
For as long as we support lambdas in GLVs with Groovy those will need to be tested indepdenently per language.
Removed test index from mergeV/E as it hasn't been maintained.
