+++ docs/src/dev/developer/for-committers.asciidoc
@@ -402,7 +402,6 @@ brackets, optionally suffixed with some additional notation depending on the pri
 * Edge - *e[_xxx_]* - The "xxx" should be replaced with a representation of an edge in the form of the
 `vertex_name-edgelabel->vertex_name`. This syntax may also include the `.id` suffix which would indicate getting the
 edge identifier or the `.sid` suffix which gets a string representation of the edge identifier.
-* Lambda - *c[_xxx_]* - The "xxx" should contain a lambda written in Groovy.
 * List - *l[_xxx_,_yyy_,_zzz_,...]* - A comma separated collection of values that make up the list should be added to
 between the square brackets. These values respect the type system thus allowing for creation of lists of vertices,
 edges, maps, and any other available type. Spaces are taken literally, therefore it is important to avoid spaces unless
@@ -581,8 +580,7 @@ the traversal does not mind the star graph limitation.
 * `@InsertionOrderingRequired` - The scenario is reliant on the graph system predictably returning results (vertices, edges, properties) in the same order in which they were inserted into the graph.
 * `@MetaProperties` - The scenario makes use of meta-properties.
 * `@MultiProperties` - The scenario makes use of multi-properties.
-* `@RemoteOnly` - The scenario uses some Gremlin syntax that cannot be supported outside of remote test executions. The
-best example of this sort of test would be one that uses the remote `Lambda` syntax.
+* `@RemoteOnly` - The scenario uses some Gremlin syntax that cannot be supported outside of remote test executions.
 * `@UserSuppliedVertexIds` - The scenario relies on the vertex IDs specified in the dataset used by the scenario.
 * `@UserSuppliedEdgeIds` - The scenario relies on the edge IDs specified in the dataset used by the scenario.
 * `@UserSuppliedVertexPropertyIds` - The scenario relies on the vertex property IDs specified in the dataset used by the scenario.
