+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/structure/Graph.java
@@ -1274,8 +1274,18 @@ public interface Graph extends AutoCloseable, Host {
         String SUITE_STRUCTURE_INTEGRATE = "org.apache.tinkerpop.gremlin.structure.StructureIntegrateSuite";
         String SUITE_PROCESS_COMPUTER = "org.apache.tinkerpop.gremlin.process.ProcessComputerSuite";
         String SUITE_PROCESS_STANDARD = "org.apache.tinkerpop.gremlin.process.ProcessStandardSuite";
+        /**
+         * @deprecated As of release 3.8.0, replaced by {@link #SUITE_PROCESS_EMBEDDED_STANDARD}.
+         */
+        @Deprecated
         String SUITE_PROCESS_LIMITED_COMPUTER = "org.apache.tinkerpop.gremlin.process.ProcessLimitedComputerSuite";
+        /**
+         * @deprecated As of release 3.8.0, replaced by {@link #SUITE_PROCESS_EMBEDDED_COMPUTER}
+         */
+        @Deprecated
         String SUITE_PROCESS_LIMITED_STANDARD = "org.apache.tinkerpop.gremlin.process.ProcessLimitedStandardSuite";
+        String SUITE_PROCESS_EMBEDDED_COMPUTER = "org.apache.tinkerpop.gremlin.process.ProcessEmbeddedComputerSuite";
+        String SUITE_PROCESS_EMBEDDED_STANDARD = "org.apache.tinkerpop.gremlin.process.ProcessEmbeddedStandardSuite";
 
         /**
          * The test suite class to opt in to.
