+++ gremlin-dotnet/test/Gremlin.Net.IntegrationTest/Gherkin/CommonSteps.cs
@@ -73,7 +73,6 @@ namespace Gremlin.Net.IntegrationTest.Gherkin
                 {@"l\[(.*)\]", ToList},
                 {@"s\[(.*)\]", ToSet},
                 {@"m\[(.+)\]", ToMap},
-                {@"c\[(.+)\]", ToLambda},
                 {@"t\[(.+)\]", ToT},
                 {"null", (_, __) => null},
                 {"true", (_, __) => true},
@@ -366,11 +365,6 @@ namespace Gremlin.Net.IntegrationTest.Gherkin
         {
             var jsonMap = JsonSerializer.Deserialize<JsonElement>(stringMap, JsonDeserializingOptions);
             return ParseMapValue(jsonMap, graphName)!;
-        }
-
-        private static object ToLambda(string stringLambda, string graphName)
-        {
-            return Lambda.Groovy(stringLambda);
         }
 
         private static object ToT(string enumName, string graphName)
