+++ gremlin-python/src/main/python/radish/feature_steps.py
@@ -36,6 +36,8 @@ project = __.project
 tail = __.tail
 
 ignores = [
+    "g.V().choose(has(T.label, \"person\"),values(\"age\").groupCount(\"a\"), values(\"name\").groupCount(\"b\")).cap(\"a\", \"b\").unfold()", # uses embedded Map in assertion
+    "g.withSideEffect(\"x\",{}).V().both().both().sideEffect(__.store(\"x\").by(\"name\")).cap(\"x\").unfold()", # Objects must be both of Map or Collection: a=LinkedHashMap b=BulkSet???
     "g.withoutStrategies(CountStrategy).V().count()" # serialization issues with Class in GraphSON
     "g.withoutStrategies(LazyBarrierStrategy).V().as(\"label\").aggregate(local,\"x\").select(\"x\").select(\"label\")",
     "g.withSack(xx1, Operator.assign).V().local(__.out(\"knows\").barrier(Barrier.normSack)).in(\"knows\").barrier().sack()", # issues with BigInteger/BigDecimal - why do we carry BigDecimal? just use python Decimal module?
@@ -45,9 +47,7 @@ ignores = [
 
 @given("the {graph_name:w} graph")
 def choose_graph(step, graph_name):
-    # if we have no traversals then we are ignoring the test - should be temporary until we can settle out the
-    # issue of handling the removal of lambdas from Gremlin as a language
-    step.context.ignore = len(step.context.traversals) == 0
+    step.context.ignore = False
     tagset = [tag.name for tag in step.all_tags]
     if not step.context.ignore:
         step.context.ignore = "AllowNullPropertyValues" in tagset
@@ -272,8 +272,6 @@ def _convert(val, ctx):
     elif isinstance(val, str) and re.match(r"^p\[.*\]$", val):  # parse path
         path_objects = list(map((lambda x: _convert(x, ctx)), val[2:-1].split(",")))
         return Path([set([])], path_objects)
-    elif isinstance(val, str) and re.match(r"^c\[.*\]$", val):  # parse lambda/closure
-        return lambda: (val[2:-1], "gremlin-groovy")
     elif isinstance(val, str) and re.match(r"^t\[.*\]$", val):  # parse instance of T enum
         return T[val[2:-1]]
     elif isinstance(val, str) and re.match(r"^D\[.*\]$", val):  # parse instance of Direction enum
