+++ gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/driver/remote/AbstractRemoteGraphProvider.java
@@ -55,92 +55,12 @@ import static org.apache.tinkerpop.gremlin.process.remote.RemoteConnection.GREML
         method = "*",
         reason = "Tests for profile() are not supported for remotes")
 @Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.branch.BranchTest",
-        method = "g_V_branchXlabel_eq_person__a_bX_optionXa__ageX_optionXb__langX_optionXb__nameX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.branch.ChooseTest",
-        method = "g_V_chooseXlabel_eqXpersonX__outXknowsX__inXcreatedXX_name",
+        test = "org.apache.tinkerpop.gremlin.process.traversal.step.LambdaStepTest",
+        method = "*",
         reason = "Tests that include lambdas are not supported by the test suite for remotes")
 @Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.DedupTest",
         method = "g_V_both_name_order_byXa_bX_dedup_value",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_E_filterXfalseX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_E_filterXtrueX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_V_filterXfalseX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_V_filterXlang_eq_javaX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_V_filterXname_startsWith_m_OR_name_startsWith_pX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_V_filterXtrueX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_VX1X_filterXage_gt_30X",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_VX1X_out_filterXage_gt_30X",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.FilterTest",
-        method = "g_VX2X_filterXage_gt_30X",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.GroupTest",
-        method = "g_V_group_byXname_substring_1X_byXconstantX1XX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.GroupTest",
-        method = "g_V_groupXaX_byXname_substring_1X_byXconstantX1XX_capXaX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.filter.HasTest",
-        method = "g_V_outXcreatedX_hasXname__mapXlengthX_isXgtX3XXX_name",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.InjectTest",
-        method = "g_VX1X_out_name_injectXdanielX_asXaX_mapXlengthX_path",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MapTest",
-        method = "g_VX1X_mapXnameX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MapTest",
-        method = "g_VX1X_out_mapXnameX_mapXlengthX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MapTest",
-        method = "g_VX1X_outE_label_mapXlengthX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MapTest",
-        method = "g_withPath_V_asXaX_out_mapXa_nameX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MapTest",
-        method = "g_withPath_V_asXaX_out_out_mapXa_name_it_nameX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.OrderTest",
-        method = "g_V_hasLabelXpersonX_order_byXvalueXageX_descX_name",
         reason = "Tests that include lambdas are not supported by the test suite for remotes")
 @Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.OrderTest",
@@ -151,44 +71,8 @@ import static org.apache.tinkerpop.gremlin.process.remote.RemoteConnection.GREML
         method = "g_V_order_byXname_a1_b1X_byXname_b2_a2X_name",
         reason = "Tests that include lambdas are not supported by the test suite for remotes")
 @Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.OrderTest",
-        method = "g_VX1X_hasXlabel_personX_mapXmapXint_ageXX_orderXlocalX_byXvalues_descX_byXkeys_ascX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.branch.RepeatTest",
-        method = "g_V_repeatXbothX_untilXname_eq_marko_or_loops_gt_1X_groupCount_byXnameX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SackTest",
         method = "g_withSackXmap__map_cloneX_V_out_out_sackXmap_a_nameX_sack",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_VX1X_out_sideEffectXincr_cX_name",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_VX1X_out_sideEffectXX_name",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_VX1X_sideEffectXstore_aX_name",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_withSideEffectsXa__linkedhashmapX_withSideEffectXb__arraylist__addAllX_withSideEffectXc__arrayList__addAllX_V_groupXaX_byXlabelX_byXcountX_sideEffectXb__1_2_3X_out_out_out_sideEffectXc__bob_danielX_capXaX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_withSideEffectXa_0_sumX_V_out_sideEffectXsideEffectsXa_bulkXX_capXaX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_withSideEffectXa_0X_V_out_sideEffectXsideEffectsXa_1XX_capXaX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SideEffectTest",
-        method = "g_withSideEffectXa__linkedhashmapX_V_out_groupCountXaX_byXlabelX_out_out_capXaX",
         reason = "Tests that include lambdas are not supported by the test suite for remotes")
 @Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SubgraphTest",
@@ -201,10 +85,6 @@ import static org.apache.tinkerpop.gremlin.process.remote.RemoteConnection.GREML
 @Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.sideEffect.SubgraphTest",
         method = "g_withSideEffectXsgX_V_hasXname_danielXout_capXsgX",
-        reason = "Tests that include lambdas are not supported by the test suite for remotes")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.UnfoldTest",
-        method = "g_V_valueMap_unfold_mapXkeyX",
         reason = "Tests that include lambdas are not supported by the test suite for remotes")
 @Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.OrderabilityTest",
