+++ gremlin-test/src/main/java/org/apache/tinkerpop/gremlin/process/ProcessComputerSuite.java
@@ -27,6 +27,7 @@ import org.apache.tinkerpop.gremlin.process.computer.search.path.ShortestPathVer
 import org.apache.tinkerpop.gremlin.process.traversal.TraversalEngine;
 import org.apache.tinkerpop.gremlin.process.traversal.TraversalInterruptionComputerTest;
 import org.apache.tinkerpop.gremlin.process.traversal.step.ComplexTest;
+import org.apache.tinkerpop.gremlin.process.traversal.step.LambdaStepTest;
 import org.apache.tinkerpop.gremlin.process.traversal.step.branch.BranchTest;
 import org.apache.tinkerpop.gremlin.process.traversal.step.branch.ChooseTest;
 import org.apache.tinkerpop.gremlin.process.traversal.step.branch.LocalTest;
@@ -120,6 +121,8 @@ public class ProcessComputerSuite extends AbstractGremlinSuite {
 
             // computer, vertex program, and map/reduce semantics
             GraphComputerTest.class,
+
+            LambdaStepTest.Traversals.class,
 
             // branch
             BranchTest.Traversals.class,
@@ -219,6 +222,8 @@ public class ProcessComputerSuite extends AbstractGremlinSuite {
      * A list of the minimum set of base tests that Gremlin flavors should implement to be compliant with Gremlin.
      */
     private static final Class<?>[] testsToEnforce = new Class<?>[]{
+            LambdaStepTest.class,
+
             // branch
             BranchTest.class,
             ChooseTest.class,
