+++ gremlin-test/src/main/java/org/apache/tinkerpop/gremlin/process/traversal/step/map/UnfoldTest.java
@@ -29,8 +29,6 @@ import org.junit.runner.RunWith;
 
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import static org.apache.tinkerpop.gremlin.LoadGraphWith.GraphData.MODERN;
@@ -39,7 +37,6 @@ import static org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__.hasId;
 import static org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__.outE;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.fail;
 
 /**
  * @author Marko A. Rodriguez (http://markorodriguez.com)
@@ -48,8 +45,6 @@ import static org.junit.Assert.fail;
 public abstract class UnfoldTest extends AbstractGremlinProcessTest {
 
     public abstract Traversal<Vertex, Edge> get_g_V_localXoutE_foldX_unfold();
-
-    public abstract Traversal<Vertex, String> get_g_V_valueMap_unfold_mapXkeyX();
 
     public abstract Traversal<Vertex, String> get_g_VX1X_repeatXboth_simplePathX_untilXhasIdX6XX_path_byXnameX_unfold(final Object v1Id, final Object v6Id);
 
@@ -71,35 +66,6 @@ public abstract class UnfoldTest extends AbstractGremlinProcessTest {
 
     @Test
     @LoadGraphWith(MODERN)
-    public void g_V_valueMap_unfold_mapXkeyX() {
-        final Traversal<Vertex, String> traversal = get_g_V_valueMap_unfold_mapXkeyX();
-        printTraversalForm(traversal);
-        int counter = 0;
-        int ageCounter = 0;
-        int nameCounter = 0;
-        int langCounter = 0;
-        while (traversal.hasNext()) {
-            counter++;
-            final String key = traversal.next();
-            if (key.equals("name"))
-                nameCounter++;
-            else if (key.equals("age"))
-                ageCounter++;
-            else if (key.equals("lang"))
-                langCounter++;
-            else
-                fail("The provided key is not known: " + key);
-        }
-        assertEquals(12, counter);
-        assertEquals(4, ageCounter);
-        assertEquals(2, langCounter);
-        assertEquals(6, nameCounter);
-        assertEquals(counter, ageCounter + langCounter + nameCounter);
-        assertFalse(traversal.hasNext());
-    }
-
-    @Test
-    @LoadGraphWith(MODERN)
     public void g_VX1X_repeatXboth_simplePathX_untilXhasIdX6XX_path_byXnameX_unfold() {
         final Traversal<Vertex, String> traversal = get_g_VX1X_repeatXboth_simplePathX_untilXhasIdX6XX_path_byXnameX_unfold(
                 convertToVertexId("marko"), convertToVertexId("peter"));
@@ -112,11 +78,6 @@ public abstract class UnfoldTest extends AbstractGremlinProcessTest {
         @Override
         public Traversal<Vertex, Edge> get_g_V_localXoutE_foldX_unfold() {
             return g.V().local(outE().fold()).unfold();
-        }
-
-        @Override
-        public Traversal<Vertex, String> get_g_V_valueMap_unfold_mapXkeyX() {
-            return g.V().valueMap().<Map.Entry<String, List>>unfold().map(m -> m.get().getKey());
         }
 
         @Override
