+++ gremlin-test/src/main/resources/org/apache/tinkerpop/gremlin/test/features/branch/Repeat.feature
@@ -212,18 +212,6 @@ Feature: Step - repeat()
       | ripple  |
       | lop  |
 
-  Scenario: g_V_repeatXbothX_untilXname_eq_marko_or_loops_gt_1X_groupCount_byXnameX
-    Given the modern graph
-    And using the parameter pred1 defined as "c[t -> t.get().value('name').equals('lop') || t.loops() > 1]"
-    And the traversal of
-      """
-      g.V().repeat(__.both()).until(pred1).groupCount().by("name")
-      """
-    When iterated to list
-    Then the result should be unordered
-      | result |
-      | m[{"ripple":"d[3].l","vadas":"d[3].l","josh":"d[4].l","lop":"d[10].l","marko":"d[4].l"}] |
-
   @GraphComputerVerificationReferenceOnly
   Scenario: g_V_hasXname_markoX_repeatXoutE_inV_simplePathX_untilXhasXname_rippleXX_path_byXnameX_byXlabelX
     Given the modern graph
