+++ gremlin-test/src/main/resources/org/apache/tinkerpop/gremlin/test/features/map/Map.feature
@@ -18,13 +18,12 @@
 @StepClassMap @StepMap
 Feature: Step - map()
 
-  Scenario: g_VX1X_mapXnameX
+  Scenario: g_VX1X_mapXvaluesXnameXX
     Given the modern graph
     And using the parameter vid1 defined as "v[marko].id"
-    And using the parameter l1 defined as "c[it.get().value('name')]"
     And the traversal of
       """
-      g.V(vid1).map(l1)
+      g.V(vid1).map(__.values("name"))
       """
     When iterated to list
     Then the result should be unordered
@@ -34,10 +33,9 @@ Feature: Step - map()
   Scenario: g_VX1X_outE_label_mapXlengthX
     Given the modern graph
     And using the parameter vid1 defined as "v[marko].id"
-    And using the parameter l1 defined as "c[it.get().length()]"
     And the traversal of
       """
-      g.V(vid1).outE().label().map(l1)
+      g.V(vid1).outE().label().map(__.length())
       """
     When iterated to list
     Then the result should be unordered
@@ -46,14 +44,12 @@ Feature: Step - map()
       | d[5].i |
       | d[5].i |
 
-  Scenario: g_VX1X_out_mapXnameX_mapXlengthX
+  Scenario: g_VX1X_out_mapXvaluesXnameXX_mapXlengthX
     Given the modern graph
     And using the parameter vid1 defined as "v[marko].id"
-    And using the parameter l1 defined as "c[it.get().value('name')]"
-    And using the parameter l2 defined as "c[it.get().toString().length()]"
     And the traversal of
       """
-      g.V(vid1).out().map(l1).map(l2)
+      g.V(vid1).out().map(__.values("name")).map(__.length())
       """
     When iterated to list
     Then the result should be unordered
@@ -62,27 +58,12 @@ Feature: Step - map()
       | d[5].i |
       | d[4].i |
 
-  @RemoteOnly
-  Scenario: g_VX1X_out_mapXlambdaXnameXX_mapXlambdaXlengthXX
+  @GraphComputerVerificationStarGraphExceeded
+  Scenario: g_withPath_V_asXaX_out_mapXselectXaX_valuesXnameXX
     Given the modern graph
-    And using the parameter vid1 defined as "v[marko].id"
     And the traversal of
       """
-      g.V(vid1).out().map(Lambda.function("it.get().value('name')")).map(Lambda.function("it.get().toString().length()"))
-      """
-    When iterated to list
-    Then the result should be unordered
-      | result |
-      | d[3].i |
-      | d[5].i |
-      | d[4].i |
-
-  Scenario: g_withPath_V_asXaX_out_mapXa_nameX
-    Given the modern graph
-    And using the parameter l1 defined as "c[it.path('a').value('name')]"
-    And the traversal of
-      """
-      g.withPath().V().as("a").out().map(l1)
+      g.withPath().V().as("a").out().map(__.select("a").values("name"))
       """
     When iterated to list
     Then the result should be unordered
@@ -94,12 +75,12 @@ Feature: Step - map()
       | josh  |
       | peter |
 
-  Scenario: g_withPath_V_asXaX_out_out_mapXa_name_it_nameX
+  @GraphComputerVerificationStarGraphExceeded
+  Scenario: g_withPath_V_asXaX_out_out_asXbX_mapXselectXaX_valuesXnameX_concatXselectXbX_valuesXnameXXX
     Given the modern graph
-    And using the parameter l1 defined as "c[it.path('a').value('name')  + it.get().value('name')]"
     And the traversal of
       """
-      g.withPath().V().as("a").out().out().map(l1)
+      g.withPath().V().as("a").out().out().as("b").map(__.select("a").values("name").concat(__.select("b").values("name")))
       """
     When iterated to list
     Then the result should be unordered
