+++ gremlin-test/src/main/resources/org/apache/tinkerpop/gremlin/test/features/map/MergeEdge.feature
@@ -18,115 +18,6 @@
 @StepClassMap @StepMergeE
 Feature: Step - mergeE()
 
-  # TEST INDEX
-  # Use of inject() is meant to exercise mergeV() in a mid-traversal form rather than start step.
-  # When the test name places a "1" following a "name" it just means that the test is using the id rather
-  # than vertex reference.
-  #
-  # g_mergeEXlabel_knows_out_marko_in_vadasX
-  # g_mergeEXlabel_knows_out_marko1_in_vadas1X
-  # g_mergeEXlabel_knows_out_marko_in_vadasX_aliased_direction
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with no option()
-  #   - vertices already exist
-  #   - results in new edge
-  # g_withSideEffectXa_label_knows_out_marko_in_vadasX_mergeEXselectXaXX
-  #   - mergeE(Traversal) specifying out/in for vertices in the match/create with no option()
-  #   - vertices already exist
-  #   - results in new edge
-  # g_mergeEXlabel_knows_out_marko_in_vadas_weight_05X_exists
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with no option()
-  #   - vertices already exist as does an edge between them but the search will not match because of a missing property
-  #   - results in new edge
-  # g_mergeEXlabel_knows_out_marko_in_vadas_weight_05X
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with no option()
-  #   - no vertices/edges in graph
-  #   - results in error as vertices don't exist
-  # g_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_created_YX_optionXonMatch_created_NX
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with option(Map)
-  #   - no vertices/edges in graph
-  #   - results in error as vertices don't exist
-  # g_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_created_YX_optionXonMatch_created_NX_exists
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with option(Map)
-  #   - vertices and edge already exist
-  #   - results in the existing edge getting a new property
-  # g_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_created_YX_optionXonMatch_created_NX_exists_updated
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with option(Map)
-  #   - vertices and edge already exist
-  #   - results in the existing edge getting an existing property updated
-  # g_V_hasXperson_name_marko_X_mergeEXlabel_knowsX_optionXonCreate_created_YX_optionXonMatch_created_NX_exists_updated
-  #   - mergeE(Map) using the vertex of the current traverser as reference and no out/in for match/create with option(Map)
-  #   - vertices and two edges of the same label already exist with one edge having an existing property
-  #   - results in one edge getting an existing property updated and the other getting a new property added
-  # g_injectXlabel_knows_out_marko_in_vadasX_mergeE
-  #   - mergeE() using Map from current traverser in the match/create with no option()
-  #   - vertices already exist
-  #   - results in a new edge
-  # g_mergeEXlabel_knows_in_vadasX_optionXonCreate_created_YX_optionXonMatch_created_NX_exists_updated
-  #   - mergeE(Map) specifying in vertex in the match/create with option(Map)
-  #   - vertices exist and there are two edges such that only one will match on the out vertex
-  #   - results in updating the property on the one matched edge
-  # g_mergeEXout_vadasX_optionXonCreate_created_YX_optionXonMatch_created_NX_exists_updated
-  #   - mergeE(Map) specifying no label and in vertex in the match/create with option(Map)
-  #   - vertices exist and there are two edges such that only one will match on the out vertex
-  #   - results in updating the property on the one matched edge
-  # g_V_hasXperson_name_marko_X_mergeEXlabel_self_out_vadas1_in_vadas1X
-  #   - mergeE(Map) specifying label and in/out vertex in the match/create with option(Map)
-  #   - vertices exist with no edge and the match/create map overrides the marko vertex in the traverser
-  #   - results in a new edge
-  # g_withSideEffectXc_created_YX_withSideEffectXm_matchedX_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_selectXcXX_optionXonMatch_selectXmXX
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with option(Map)
-  #   - vertices exist in the graph
-  #   - results in one new edge
-  # g_withSideEffectXc_created_YX_withSideEffectXm_matchedX_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_selectXcXX_optionXonMatch_selectXmXX_exists
-  #   - mergeE(Map) specifying out/in for vertices in the match/create with option(Map)
-  #   - vertices and edge already exist
-  #   - results in the existing edge getting a new property
-  # g_injectXlabel_knows_out_marko_in_vadas_label_self_out_vadas_in_vadasX
-  #   - mergeE() using the map of current traverser as reference
-  #   - vertices already exists
-  #   - results in two new edges
-  # g_mergeEXnullX
-  # g_V_mergeEXnullX
-  #   - mergeE(null) and no option
-  #   - vertices already exists
-  #   - Directions not specified - error
-  # g_mergeEXemptyX
-  #   - mergeE(empty) and no option
-  #   - vertex present and no edges
-  #   - results in error because empty map matches all edges and there are none so it creates but doesn't have a Vertex reference to use to create the edge
-  # g_V_mergeEXemptyX
-  #   - mergeE(empty) and no option
-  #   - two vertices exist
-  #   - results in one edge added to the first vertex and then the same edge being returned
-  # g_mergeEXemptyX
-  #   - mergeE(empty) and no option
-  #   - vertices already exists
-  #   - results in no matched edges and no new edge
-  # g_mergeEXemptyX_exists
-  #   - mergeE(empty) and no option
-  #   - vertex with self-edge already exists
-  #   - results in the matched self-edge
-  # g_V_mergeEXemptyX_two_exist
-  #   - mergeE(empty) and no option
-  #   - two vertices exist
-  #   - results in two new self edges, one for each vertex
-  # g_V_mergeEXemptyX_optionXonCreate_nullX
-  #   - mergeE(empty) with onCreate(null)
-  #   - one existing vertex
-  #   - results in no new edges because onCreate() was null
-  # g_V_mergeEXlabel_selfX_optionXonMatch_nullX
-  #   - mergeE(Map) with onMatch(null)
-  #   - one existing vertex and one edge
-  #   - results in one matched edge with no updates
-  # g_V_mergeEXlabel_selfX_optionXonMatch_emptyX
-  #   - mergeE(Map) with onMatch(null)
-  #   - one existing vertex and one edge
-  #   - results in one matched edge with no updates
-  # g_withSideEffectXc_created_YX_withSideEffectXm_matchedX_mergeEXlabel_knows_out_marko_in_vadasX_optionXonCreate_selectXcXX_optionXonMatch_sideEffectXpropertiesXweightX_dropX_selectXmXX_exists
-  #   - mergeE(Map) with onMatch(Map)
-  #   - matches an edge
-  #   - results in matched edge being updated with a side-effect that drops a property that is not updated in the Map
-
   Scenario: g_V_mergeEXlabel_selfX_optionXonMatch_emptyX
     Given the empty graph
     And the graph initializer of
