+++ gremlin-test/src/main/resources/org/apache/tinkerpop/gremlin/test/features/map/MergeVertex.feature
@@ -18,90 +18,6 @@
 @StepClassMap @StepMergeV
 Feature: Step - mergeV()
 
-  # TEST INDEX
-  # Use of inject() is meant to exercise mergeV() in a mid-traversal form rather than start step
-  #
-  # g_mergeVXlabel_person_name_stephenX
-  # g_injectX0X_mergeVXlabel_person_name_stephenX
-  #   - mergeV(Map) and no option()
-  #   - results in new vertex
-  # g_mergeVXlabel_person_name_markoX
-  # g_injectX0X_mergeVXlabel_person_name_markoX
-  #   - mergeV(Map) and no option()
-  #   - results in found vertex
-  # g_mergeVXlabel_person_name_stephenX_optionXonCreate_label_person_name_stephen_age_19X_option
-  # g_injectX0X_mergeVXlabel_person_name_stephenX_optionXonCreate_label_person_name_stephen_age_19X_option
-  #   - mergeV(Map) using onCreate(Map)
-  #   - results in new vertex
-  # g_mergeVXlabel_person_name_markoX_optionXonMatch_age_19X_option
-  # g_injectX0X_mergeVXlabel_person_name_markoX_optionXonMatch_age_19X_option
-  #   - mergeV(Map) using onMatch(Map)
-  #   - results in update vertex
-  # g_withSideEffectXc_label_person_name_stephenX_withSideEffectXm_label_person_name_stephen_age_19X_mergeVXselectXcXX_optionXonCreate_selectXmXX_option
-  # g_withSideEffectXc_label_person_name_stephenX_withSideEffectXm_label_person_name_stephen_age_19X_injectX0X_mergeVXselectXcXX_optionXonCreate_selectXmXX_option
-  #   - mergeV(Traversal) grabbing side-effect Map with onCreate(Traversal)
-  #   - results in new vertex
-  # g_withSideEffectXc_label_person_name_markoX_withSideEffectXm_age_19X_mergeVXselectXcXX_optionXonMatch_selectXmXX_option
-  # g_withSideEffectXc_label_person_name_markoX_withSideEffectXm_age_19X_injectX0X_mergeVXselectXcXX_optionXonMatch_selectXmXX_option
-  #   - mergeV(Traversal) grabbing side-effect Map with onMatch(Traversal)
-  #   - results in updated vertex
-  # g_mergeVXlabel_person_name_markoX_propertyXname_vadas_acl_publicX
-  # g_injectX0X_mergeVXlabel_person_name_markoX_propertyXname_vadas_acl_publicX
-  #   - mergeV(Map) with no option and call to AddPropertyStep
-  #   - results in updated vertex and added meta-property
-  # g_injectXlabel_person_name_marko_label_person_name_stephenX_mergeVXidentityX
-  #   - mergeV(Traversal) grabbing current Map from traverser for the search criteria with no option()
-  #   - result in one found vertex and one new vertex
-  # g_injectXlabel_person_name_marko_label_person_name_stephenX_mergeV
-  #   - mergeV() which assumes incoming Map on the traverser for the search criteria with no option()
-  #   - results in one found vertex and one new vertex
-  # g_mergeVXlabel_person_name_stephenX_propertyXlist_name_steveX
-  #   - mergeV() which assumes finding vertex with list cardinality property and call AddPropertyStep
-  #   - results in updated vertex with additional list cardinality property
-  # g_mergeXlabel_person_name_vadasX_optionXonMatch_age_35X
-  #   - mergeV(Map) using onMatch(Map)
-  #   - results in updating two matched vertices
-  # g_V_mapXmergeXlabel_person_name_joshXX
-  #   - mergeV(Map) with no option() - testing child traversal usage
-  #   - results in one new vertex and one existing vertex that was just created
-  # g_mergeVXnullX
-  # g_V_mergeVXnullX
-  #   - mergeV(null) with no option()
-  #   - results in no new vertex and nothing returned
-  # g_mergeVXemptyX
-  # g_V_mergeVXemptyX
-  #   - mergeV(empty) with no option()
-  #   - results in matched vertex with no updates
-  # g_mergeVXemptyX_no_existing
-  # g_injectX0X_mergeVXemptyX_no_existing
-  #   - mergeV(empty) with no option()
-  #   - results in not matched updates and a creation of a vertex with default values
-  # g_mergeVXnullX_optionXonCreate_emptyX
-  # g_V_mergeVXnullX_optionXonCreate_emptyX
-  #   - mergeV(null) with onCreate(empty)
-  #   - results in no matches and creates a default vertex
-  # g_mergeVXlabel_person_name_stephenX_optionXonCreate_nullX
-  # g_V_mergeVXlabel_person_name_stephenX_optionXonCreate_nullX
-  #   - mergeV(Map) with onCreate(null)
-  #   - results in no match and no vertex creation
-  # g_mergeVXnullX_optionXonCreate_label_null_name_markoX
-  # g_V_mergeVXnullX_optionXonCreate_label_null_name_markoX
-  #   - mergeV(null) with onCreate(Map) where Map has a null label
-  #   - results in error
-  # g_mergeVXemptyX_optionXonMatch_nullX
-  # g_V_mergeVXemptyX_optionXonMatch_nullX
-  #   - mergeV(empty) with onMatch(null)
-  #   - results in a match and no vertex update
-  # g_V_mergeVXemptyX_two_exist
-  #  - mergeV(empty) with no option()
-  #  - results in matching two vertices
-  # g_withSideEffectXc_label_person_name_markoX_withSideEffectXm_age_19X_mergeVXselectXcXX_optionXonMatch_sideEffectXpropertiesXageX_dropX_selectXmXX_option
-  #  - mergeV() that has a match to a vertex with multiproperties
-  #  - results in all multiproperties removed to be replaced with one property
-  # g_withSideEffectXm_age_19X_V_hasXperson_name_markoX_mergeVXselectXcXX_optionXonMatch_sideEffectXpropertiesXageX_dropX_selectXmXX_option
-  #  - mergeV() that has a match to a vertex with multiproperties
-  #  - results in all multiproperties removed to be replaced with one property
-
   Scenario: g_mergeVXemptyX_optionXonMatch_nullX
     Given the empty graph
     And the graph initializer of
